Response handling for successful POSTs to /feedback
Removed success and error decoders, since the POST request only returns
a text response if it fails.

Will have to address the failure case more thoroughly in a future commit,
right now a generic error message will be displayed.

Added feedbackResult to the model and used it to conditionally render
a success/error message to the user on form submit.

Finally, redirected to the main view if feedback is successfully
submitted.