Remove success message on keyboard input
Once feedback is submitted, a "toast" of sorts is displayed to
show the user whether it succeeded or failed. This message should
not be permanent; I tried a few things, such as using Task.sequence
or andThen to clear the message after a timeout (Process.sleep),
or a similar pattern where the toast would always clear after
a page navigation event, then would be added only after the post-success
navigation command. These were complicated and I couldn't get the typing
to work out; clearing the stuff on input is simple and handles all cases.