Add click handler to links
There was a bug where ctrl-keying the submit/exit actions on the feedback
form would have the correct effect, but clicking the links would just
navigate back to /summary (without actually submitting/clearing the form).

Added a 4th argument to the footerLink function, which is passed to the
onClick handler for the links. Not crazy about putting an onClick on
an <a>, but it does the trick. "Normal" links use a NoOp because there
is no effect other than navigation; added two new Msgs, SubmitFeedback
and ExitFeedback, to handle the cases at hand.