Fix handling of absent settings fields
I'd been handling the initial load of the app (for the first time
in a given browser) by relying on a Json.Decode error, which
then provided default values.

This was a problem, since I'm now using a separate API URL flag,
which needs to be defined. So if 'theme' isn't in localStorage,
requests from the feedback form were bombing.

Used Json.Decode.oneOf to pass a default theme value correctly.